refactor(app): alterando tipo de id nos repository para seguir
o padrao, sendo integer para todos os id

+++ src/main/java/com/project/extension/repository/AtributoProdutoRepository.java
@@ -3,5 +3,5 @@ package com.project.extension.repository;
 import com.project.extension.entity.AtributoProduto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AtributoProdutoRepository extends JpaRepository<AtributoProduto, Long> {
+public interface AtributoProdutoRepository extends JpaRepository<AtributoProduto, Integer> {
 }
