+++ src/main/java/com/project/extension/repository/EstoqueRepository.java
@@ -3,5 +3,5 @@ package com.project.extension.repository;
 import com.project.extension.entity.Estoque;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstoqueRepository extends JpaRepository<Estoque, Long> {
+public interface EstoqueRepository extends JpaRepository<Estoque, Integer> {
 }
